+++ NEWS.md
@@ -1,6 +1,7 @@
 # auk 0.5.1
 
 - drop `data.table` dependency, no longer needed with `readr` speed improvements
+- fix bug arising from 'breeding bird atlas code' being renamed to 'breeding code' (issue #58)
 
 # auk 0.5.0